storefront/requirements.txt
asgiref==3.4.1
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.3.7
configobj==5.0.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.4
defusedxml==0.7.1
distro-info===0.18ubuntu0.18.04.1
Django==3.2.11
django-debug-toolbar==3.2.4
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
djoser==2.1.0
drf-nested-routers==0.93.4
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
importlib-metadata==1.7.0
incremental==16.10.1
itypes==1.2.0
Jinja2==3.0.3
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
MarkupSafe==2.0.1
mysqlclient==2.0.3
netifaces==0.10.4
oauthlib==3.1.1
PAM==0.4.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
PyGObject==3.26.1
PyJWT==1.7.1
pyOpenSSL==17.5.0
pyserial==3.4
python-apt==1.6.4
python-debian==0.1.32
python3-openid==3.2.0
pytz==2021.3
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
service-identity==16.0.0
six==1.11.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
ssh-import-id==5.7
systemd-python==234
Twisted==17.9.0
typing_extensions==4.0.1
ufw==0.36
uritemplate==4.1.1
urllib3==1.22
wrapt==1.13.3
zipp==3.6.0
zope.interface==4.3.2
